+++ src/main/java/org/parseplatform/client/Parse.java
@@ -625,6 +625,47 @@ public class Parse {
         }
     }
 
+    public static class File {
+        public static final String IMAGE_JPEG = "image/jpeg";
+        public static final String IMAGE_PNG = "image/png";
+
+        public static void upload(String base64file, String fileName, String mimeType, final AsyncCallback<ParseResponse> callback) {
+//            String data = "{__ContentType : \"" + mimeType + "\", base64 : " + base64file + "}";
+            JSONObject data = new JSONObject();
+            data.put("__ContentType", new JSONString(mimeType));
+            data.put("base64", new JSONString(base64file));
+            Shape.post(Parse.SERVER_URL + Parse.FILES_URI + fileName)
+                    .header("X-Parse-Application-Id", X_Parse_Application_Id)
+                    .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
+                    .header("X-Parse-Master-Key", X_Parse_Master_Key)
+                    .header("X-Parse-Session-Token", X_Parse_Session_Token)
+                    .header("Content-Type", mimeType)
+                    .body(data.toString())
+                    .asJson(new AsyncCallback<String>() {
+                        @Override
+                        public void onFailure(Throwable throwable) {
+                            callback.onFailure(throwable);
+                        }
+                        @Override
+                        public void onSuccess(String s) {
+                            try {
+                                JSONObject jsonObject = JSONParser.parseStrict(s).isObject();
+                                ParseResponse response = new ParseResponse();
+                                Iterator<String> it = jsonObject.keySet().iterator();
+                                while (it.hasNext()) {
+                                    String key = it.next();
+                                    JSONValue jsonValue = jsonObject.get(key);
+                                    response.put(key, jsonValue);
+                                }
+                                callback.onSuccess(response);
+                            } catch (Exception e) {
+                                callback.onFailure(e);
+                            }
+                        }
+                    });
+        }
+    }
+
     public static class Query extends JSONObject {
 
         private Where where;
@@ -825,6 +866,7 @@ public class Parse {
 
     public static String SERVER_URL = "http://localhost:1337/parse";
     public static String CLASSES_URI = "classes/";
+    public static String FILES_URI = "files/";
 
     /*
     Headers